learnt a bunch of new tags
It's interesting to see that many of these tags don't necessarily DO anything.  Just as we discussed with the span tag, they exist entirely so that A: anyone reading the html can easily differentiate all the sections or ctrl+f to find specific sections, and B: for CSS.  I can definitely imagine companies with big websites having ridiculously large css style sheets, so that the content team can just worry about wrapping everything in the right tags for brand consistency.